Change `ui.greet()` to `ui.greet($event.target.value)` in actions.mdx
The docs are incorrect! My code looked like this:

```
<input name="name" (input)="ui.greet()" />
<div>{{ ui.greet$ | async }}</div>
```

However, I got an `Expected 1 arguments, but got 0.` yet I followed the docs perfectly! Turns out, I was missing `$event.target.value`. So here's how it should've looked like in the docs:

```
<input name="name" (input)="ui.greet($event.target.value)" />
<div>{{ ui.greet$ | async }}</div>
```

And now it works. Hopefully, other devs won't be left discouraged to try this amazing tool because of this silly doc discrepancy!