feat(template): set the parent flag to false by default
BREAKING CHANGE: It is possible that this change breaks the integration with 3rd party libraries, that are not up to date. If you are using the @rx-angular/template directives to project content into other components which don't use the new signal queries, they might not work properly anymore.
In order to fix this, you might want to set the parent flag to true on a global scope by providing an `RX_RENDER_STRATEGIES_CONFIG`. Another option is to set `parent: true` on a per directive basis.
Read more about this here: https://www.rx-angular.io/docs/template/performance-issues/handling-view-and-content-queries & here: https://push-based.io/article/new-features-for-rxangular-native-signal-support-and-improved-state#the-parent-flag-gets-deprecated